#55 -  Improve Hashtag Search Functionality
Implementing a feature to manage hashtags associated with posts.

Interestingly, there is no separate method for storing hashtags. The storage of hashtags occurs only when posts are created or updated, as per the current structure of the business logic. This functionality is entirely handled within the post service through relationship mapping.

Retrieve hashtag information from the database by hashtag name.
Parse hashtags from the content of a post.
Delete hashtags that are no longer associated with any posts.
TODO: A test for the delete functionality was not created during the lecture. It would be good to add this.