#76 - AuditingFields: Modify Field Access Modifier to protected
This section should originally have been protected to align with the nature of an abstract class. However, the initial design was overly restrictive.

Now, the member entity needs to directly reference this part, so it has been properly adjusted.

With this change, member information can be stored without authentication in the constructor.

The scenarios where this is needed include member registration and member creation.